Renames progressbar to progress-bar
