+++ packages/web-components/src/index.ts
@@ -1,2 +1,2 @@
 export * from './text/index.js';
-export * from './progressbar/index.js';
+export * from './progress-bar/index.js';
