+++ packages/web-components/src/progress-bar/define.ts
@@ -1,4 +1,4 @@
 import { FluentDesignSystem } from '../fluent-design-system.js';
-import { definition } from './progressbar.definition.js';
+import { definition } from './progress-bar.definition.js';
 
 definition.define(FluentDesignSystem.registry);
