+++ packages/web-components/src/progress-bar/index.ts
@@ -1,0 +1,5 @@
+export * from './progress-bar.js';
+export * from './progress-bar.options.js';
+export { template as ProgressBarTemplate } from './progress-bar.template.js';
+export { styles as ProgressBarStyles } from './progress-bar.styles.js';
+export { definition as ProgressBarDefinition } from './progress-bar.definition.js';
