+++ packages/web-components/src/progress-bar/progress-bar.options.ts
