+++ packages/web-components/src/progress-bar/progress-bar.styles.ts
