+++ packages/web-components/src/progress-bar/progress-bar.template.ts
@@ -1,6 +1,6 @@
 import type { ElementViewTemplate } from '@microsoft/fast-element';
 import { progressTemplate } from '@microsoft/fast-foundation';
-import type { ProgressBar } from './progressbar.js';
+import type { ProgressBar } from './progress-bar.js';
 
 export const template: ElementViewTemplate<ProgressBar> = progressTemplate({
   indeterminateIndicator1: `<span class="indeterminate-indicator-1" part="indeterminate-indicator-1></span>`,
