--- packages/web-components/src/progressbar/index.ts
@@ -1,5 +1,0 @@
-export * from './progressbar.js';
-export * from './progressbar.options.js';
-export { template as ProgressBarTemplate } from './progressbar.template.js';
-export { styles as ProgressBarStyles } from './progressbar.styles.js';
-export { definition as ProgressBarDefinition } from './progressbar.definition.js';
